+++ R/data.R
@@ -8,3 +8,14 @@
 #'
 #' @references Lumley T (2010). Complex Surveys: a guide to analysis using R. Wiley
 NULL
+
+
+#' @docType data
+#' @title Sample dataset from the EFC Survey
+#' @name efc
+#' @keywords data
+#'
+#' @description Selected variables from the EUROFAMCARE survey. Useful when
+#' testing on "real-life" data sets, including random missing values. This
+#' data set also has value and variable label attributes.
+NULL