+++ tests/testthat/test-convert_data_to_numeric.R
@@ -33,3 +33,52 @@ test_that("convert factor to numeric, dummy factors", {
     ignore_attr = TRUE
   )
 })
+
+test_that("convert factor to numeric, dummy factors, with NA", {
+  x1 <- factor(rep(c("a", "b"), 3))
+  x2 <- factor(c("a", NA_character_, "a", "b", "a", "b"))
+  x3 <- factor(c(NA_character_, "b", "a", "b", "a", "b"))
+  x4 <- factor(c("a", "b", "a", "b", "a",  NA_character_))
+  x5 <- factor(c(NA_character_, "b", "a", "b", "a", NA_character_))
+  x6 <- factor(c(NA_character_, "b", NA_character_, "b", "a", NA_character_))
+  x7 <- factor(c(NA_character_, "b", "a", "b", "a", "b", NA_character_, "b", "a", NA_character_, "a", "b", "a", "b", "a", NA_character_))
+
+  # same observations are missing
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x1, dummy_factors = TRUE))),
+    which(is.na(x1))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x2, dummy_factors = TRUE))),
+    which(is.na(x2))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x3, dummy_factors = TRUE))),
+    which(is.na(x3))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x4, dummy_factors = TRUE))),
+    which(is.na(x4))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x5, dummy_factors = TRUE))),
+    which(is.na(x5))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x6, dummy_factors = TRUE))),
+    which(is.na(x6))
+  )
+  expect_equal(
+    which(!complete.cases(convert_data_to_numeric(x7, dummy_factors = TRUE))),
+    which(is.na(x7))
+  )
+
+  # output has same number of observation as input
+  expect_equal(nrow(convert_data_to_numeric(x1, dummy_factors = TRUE)), length(x1))
+  expect_equal(nrow(convert_data_to_numeric(x2, dummy_factors = TRUE)), length(x2))
+  expect_equal(nrow(convert_data_to_numeric(x3, dummy_factors = TRUE)), length(x3))
+  expect_equal(nrow(convert_data_to_numeric(x4, dummy_factors = TRUE)), length(x4))
+  expect_equal(nrow(convert_data_to_numeric(x5, dummy_factors = TRUE)), length(x5))
+  expect_equal(nrow(convert_data_to_numeric(x6, dummy_factors = TRUE)), length(x6))
+  expect_equal(nrow(convert_data_to_numeric(x7, dummy_factors = TRUE)), length(x7))
+})