+++ tests/testthat/test-data_match.R
@@ -9,3 +9,25 @@ test_that("data_match works as expected", {
   expect_equal(unique(df2$vs), 0)
   expect_equal(unique(df2$am), c(1, 0))
 })
+
+
+if (require("poorman")) {
+  data(efc, package = "datawizard")
+  test_that("data_match works with missing data", {
+    # "OR" works
+    x1 <- length(data_match(efc, data.frame(c172code = 1, e16sex = 2), match = "or"))
+    x2 <- nrow(poorman::filter(efc, c172code == 1 | e16sex == 2))
+    expect_equal(x1, x2)
+
+    # "AND" works
+    x1 <- length(data_match(efc, data.frame(c172code = 1, e16sex = 2), match = "and"))
+    x2 <- nrow(poorman::filter(efc, c172code == 1, e16sex == 2))
+    expect_equal(x1, x2)
+
+    # "NOT" works
+    x1 <- length(data_match(efc, data.frame(c172code = 1, e16sex = 2), match = "not"))
+    x2 <- nrow(poorman::filter(efc, c172code != 1, e16sex != 2))
+    expect_equal(x1, x2)
+  })
+}
+